Add conftest file for tasks tests
